cache burst apm url and measure tr and spans
